requirements.txt
fastapi
yolov5
Unidecode
google-cloud
google-cloud-vision